+++ changelog.md
@@ -1,4 +1,7 @@
 # Changelog
 
-**./test-parts/part-c-internal_scripts/scripts/s-request_api-other.js**
-* Renamed 'checkRandomIp' function to 'checkRandomIpAddress'
+**./test-parts/part-b-external_modules/modules/pkg-validator.js**
+* Renamed functions:
+	* 'verifyIpAddressVerFour' to 'verifyIpAddressFour'
+	* 'verifyIpAddressVerSix' to 'verifyIpAddressSix'
+	* 'verifyMacAddressFunction' to 'verifyMacAddress'